Fix: Change nextauth execution logic
